PulseAudio: Rename internal headers to have '_internal.h' file-ending and make changes to CMake-files and configure.in
